refactor: 패키지명 수정, test 폴더 패키지 정리
