+++ src/test/java/calculator/CalculatorTest.java
@@ -1,6 +1,5 @@
-package calculatorTest;
+package calculator;
 
-import calculator.Calculator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
