+++ src/test/java/racingcar/Domain/CarTest.java
@@ -1,4 +1,4 @@
-package racingcarTest;
+package racingcar.Domain;
 
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
@@ -7,8 +7,6 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
 import static org.assertj.core.api.Assertions.assertThat;
-
-import racingcar.Model.Car;
 
 /**
  * 클래스 이름 : CarTest.java
